feat(.config/git/.gitignore-sync,wsl/install.sh): add global gitignore file (#123)
* feat(.config/git/.gitignore-sync,wsl/install.sh): add global gitignore file

* fix(wsl/install.sh): fix find command arguments

+++ wsl/install.sh
@@ -37,7 +37,7 @@ git checkout "${git_ref}"
 wsl_dir="$(realpath ./wsl)"
 cd "${wsl_dir}" || exit
 
-paths="$(find . -type f ! -name "install.sh" ! -name "setup-git.sh")"
+paths="$(find . -type f ! -name "install.sh" ! -name "setup-git.sh" ! -name ".gitignore-sync")"
 for path in ${paths}; do
 	# append to .bashrc or .profile because WSL has default .bashrc and .profile
 	if [[ -f ${path} && ${path#./} == ".bashrc" ]]; then
@@ -53,7 +53,13 @@ for path in ${paths}; do
 		continue
 	fi
 	mkdir --parents "$(dirname "${HOME}/${path#./}")"
-	ln --symbolic --no-dereference --force "${wsl_dir}/${path#./}" "${HOME}/${path#./}"
+	if [[ -f ${path} && ${path#./} == ".config/git/.gitignore" ]]; then
+		# ignore-sync doesn't support filename `ignore-sync`, so rename generated .gitignore to ignore
+		ln --symbolic --no-dereference --force "${wsl_dir}/${path#./}" "${HOME}/.config/git/ignore"
+		continue
+	else
+		ln --symbolic --no-dereference --force "${wsl_dir}/${path#./}" "${HOME}/${path#./}"
+	fi
 	echo installed "${path}"
 done
